Resize terminal when starting the game
Also changed game_win/game_lose to use @screen and allowed
the player to type 'exit' to quit the game early.

This looks much better when playing now, and even when the game ends it
is nicely lined up. Only the error messages look a little funny. Still
need to move all the UI logic to Display class.